refactor(#792): extract controller_taps DI-loop tests into their own file (sweep)
controller_taps.rs 621 -> 385 (leaves the debt list). Moved the two inline
test modules to a sibling wired via #[path]; super:: -> super::super:: for the
parent's items. infra-cpal tests pass, zero warnings.
